Adjust only transitive dependencies in the fallback
when a package from an allowed origin can't be marked to install/upgrade.

This is a much lighter approach than marking every upgradable package because
the full fallback was triggered on packages held back as well, using an
excessive amount of CPU time.

Also it crashed with packages not having any version in allowed origins.

LP: #1824804
LP: #1824949